+++ IPlayer.h
@@ -16,7 +16,7 @@ public:
 	virtual int GetPositionX() = 0;
 	virtual int GetPositionY() = 0;
   virtual int GetAttackDamage() = 0;
-  virtual float GetCurrOxyLevel() = 0;
+  virtual double GetCurrOxyLevel() = 0;
   virtual int GetCurrHP() = 0;
   virtual int GetCurrScore() = 0;
 
@@ -25,21 +25,21 @@ public:
 
   virtual void SetAttackDamage(int) = 0;
 
-  virtual void SetMaxOxygenLevel(float) = 0;
-  virtual void SetOxygenLevel(float) = 0;
+  virtual void SetMaxOxygenLevel(double) = 0;
+  virtual void SetOxygenLevel(double) = 0;
 
   virtual void SetMaxHP(int) = 0;
   virtual void SetHP(int) = 0;
 
-  virtual void SetLasgulaDuration(float) = 0;
-  virtual void SetComboDuration(float) = 0;
+  virtual void SetLasgulaDuration(double) = 0;
+  virtual void SetComboDuration(double) = 0;
 
   /* Modifiers */
   virtual void AddHP(int hp) = 0;
 
-  virtual void AddOxygen(float amountOfOxygen) = 0;
-  virtual void AddOxygenFromOxyBlock(float amountOfOxygen) = 0;
-  virtual void AddOxygenSpecialCase(float amountOfOxygen) = 0;
+  virtual void AddOxygen(double amountOfOxygen) = 0;
+  virtual void AddOxygenFromOxyBlock(double amountOfOxygen) = 0;
+  virtual void AddOxygenSpecialCase(double amountOfOxygen) = 0;
 
   virtual void AddScore(int score) = 0;
   virtual void AddScoreFromOxyBlock(int score) = 0;
