+++ PlayerDecorator.h
@@ -13,7 +13,7 @@ public:
 	virtual int GetPositionX() override { return _player->GetPositionX(); }
 	virtual int GetPositionY() override { return _player->GetPositionY(); }
   virtual int GetAttackDamage() override { return _player->GetAttackDamage(); }
-  virtual float GetCurrOxyLevel() override { return _player->GetCurrOxyLevel(); }
+  virtual double GetCurrOxyLevel() override { return _player->GetCurrOxyLevel(); }
   virtual int GetCurrHP() override { return _player->GetCurrHP(); }
   virtual int GetCurrScore() override { return _player->GetCurrScore(); }
 
@@ -22,28 +22,28 @@ public:
   virtual void SetAttackDamage(int ad) override 
   { _player->SetAttackDamage(ad); };
 
-  virtual void SetMaxOxygenLevel(float maxOxygenLevel) override 
+  virtual void SetMaxOxygenLevel(double maxOxygenLevel) override 
   { _player->SetMaxOxygenLevel(maxOxygenLevel); }
-  virtual void SetOxygenLevel(float oxygenLevel) override 
+  virtual void SetOxygenLevel(double oxygenLevel) override 
   { _player->SetOxygenLevel(oxygenLevel); }
 
   virtual void SetMaxHP(int maxHP) override 
   { _player->SetMaxHP(maxHP); }
   virtual void SetHP(int hp) override { _player->SetHP(hp); }
 
-  virtual void SetLasgulaDuration(float lasgulaDuration) 
+  virtual void SetLasgulaDuration(double lasgulaDuration) 
   { _player->SetLasgulaDuration(lasgulaDuration); }
-  virtual void SetComboDuration(float comboDuration) 
+  virtual void SetComboDuration(double comboDuration) 
   { _player->SetComboDuration(comboDuration); }
 
   /* Modifiers */
   virtual void AddHP(int hp) override { _player->AddHP(hp); }
 
-  virtual void AddOxygen(float amountOfOxygen) override 
+  virtual void AddOxygen(double amountOfOxygen) override 
   { _player->AddOxygen(amountOfOxygen); }
-  virtual void AddOxygenFromOxyBlock(float amountOfOxygen) override 
+  virtual void AddOxygenFromOxyBlock(double amountOfOxygen) override 
   { _player->AddOxygenFromOxyBlock(amountOfOxygen); };
-  virtual void AddOxygenSpecialCase(float amountOfOxygen) override 
+  virtual void AddOxygenSpecialCase(double amountOfOxygen) override 
   { _player->AddOxygenSpecialCase(amountOfOxygen); }
 
   virtual void AddScore(int score) override
